+++ src/main/java/com/github/deeepamin/ciaid/refactor/actions/GitLabCIYamlExtractReferenceAction.java
@@ -1,20 +1,21 @@
 package com.github.deeepamin.ciaid.refactor.actions;
 
-import com.intellij.ide.DataManager;
+import com.github.deeepamin.ciaid.refactor.dialogs.ExtractReferenceDialog;
+import com.github.deeepamin.ciaid.services.CIAidProjectService;
+import com.github.deeepamin.ciaid.utils.FileUtils;
+import com.github.deeepamin.ciaid.utils.YamlUtils;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
-import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.command.WriteCommandAction;
-import com.intellij.openapi.editor.Editor;
-import com.intellij.openapi.editor.ScrollType;
-import com.intellij.psi.PsiDocumentManager;
+import com.intellij.openapi.fileEditor.FileEditorManager;
 import com.intellij.psi.PsiElement;
-import com.intellij.refactoring.rename.RenameHandlerRegistry;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.yaml.YAMLElementGenerator;
+import org.jetbrains.yaml.YAMLUtil;
 import org.jetbrains.yaml.psi.YAMLFile;
 import org.jetbrains.yaml.psi.YAMLKeyValue;
-import org.jetbrains.yaml.psi.YAMLMapping;
+
+import static com.github.deeepamin.ciaid.model.gitlab.GitlabCIYamlKeywords.*;
 
 public class GitLabCIYamlExtractReferenceAction extends BaseRefactorAction {
   @Override
@@ -22,31 +23,45 @@ public class GitLabCIYamlExtractReferenceAction extends BaseRefactorAction {
     var project = e.getProject();
     var editor = e.getData(CommonDataKeys.EDITOR);
     var element = e.getData(CommonDataKeys.PSI_ELEMENT);
-    if (project == null || editor == null || element == null) {
-      // TODO show error message
+    if (project == null || editor == null || !(element instanceof YAMLKeyValue keyValue)) {
       return;
     }
+    var keyName = keyValue.getKeyText();
+    var ciAidProjectService = CIAidProjectService.getCIAidProjectService(element);
+    var files = ciAidProjectService.getPluginData()
+            .keySet()
+            .stream()
+            .toList();
+    var extractRefDialog = new ExtractReferenceDialog(project, keyName, files);
+    if (extractRefDialog.showAndGet()) {
+      extractReference(element, keyValue, extractRefDialog);
+    }
+  }
 
-    var extractedReference = WriteCommandAction.writeCommandAction(project)
-            .compute(() -> {
-      var keyValue = (YAMLKeyValue) element;
-      if (keyValue.getValue() == null) {
-        return null;
-      }
+  private void extractReference(PsiElement element, YAMLKeyValue keyValue, ExtractReferenceDialog extractRefDialog) {
+    if (element == null || keyValue.getValue() == null) {
+      return;
+    }
+    var project = element.getProject();
+    var selectedFilePath = extractRefDialog.getSelectedFilePath();
+    var virtualFile = FileUtils.getVirtualFile(selectedFilePath, project).orElse(null);
+    if (virtualFile == null) {
+      return;
+    }
+    var rootMapping = YamlUtils.getRootMapping(project, selectedFilePath);
+    if (rootMapping == null) {
+      return;
+    }
+    WriteCommandAction.runWriteCommandAction(project, () -> {
       var keyName = keyValue.getKeyText();
+      var referenceKey = extractRefDialog.getReferenceKey();
       var valueText = keyValue.getValue().getText();
-      var referenceName = ".extracted";
-      var file = (YAMLFile) element.getContainingFile();
 
-      var topLevelValue = file.getDocuments().getFirst().getTopLevelValue();
-      if (!(topLevelValue instanceof YAMLMapping rootMapping)) {
-        return null;
-      }
       var generator = YAMLElementGenerator.getInstance(project);
 
       var extractedKeyValue = generator.createYamlKeyValue(keyName, valueText);
       var extractedMapping = createYamlMapping(project, extractedKeyValue);
-      var referenceKeyValue = generator.createYamlKeyValue(referenceName, extractedMapping.getText());
+      var referenceKeyValue = generator.createYamlKeyValue(referenceKey, extractedMapping.getText());
 
       var lastChild = rootMapping.getLastChild();
       var newLine = YAMLElementGenerator.getInstance(project).createEol();
@@ -54,42 +69,37 @@ public class GitLabCIYamlExtractReferenceAction extends BaseRefactorAction {
       if (lastChild != null) {
         rootMapping.addAfter(newLine, lastChild);
       }
+      rootMapping.add(referenceKeyValue);
 
-      var extracted = (YAMLKeyValue) rootMapping.add(referenceKeyValue);
-
-      var replacement = createYamlKeyValue(project, keyName, "!reference [ " + referenceName + ", " + keyName + " ]");
+      var replacement = createYamlKeyValue(project, keyName, REFERENCE_TAG + " [ " + referenceKey + ", " + keyName + " ]");
       if (replacement != null) {
         keyValue.replace(replacement);
       }
 
-      PsiDocumentManager.getInstance(project).doPostponedOperationsAndUnblockDocument(editor.getDocument());
-
-      var extractedKey = extracted.getKey();
-      if (extractedKey != null) {
-        editor.getCaretModel().moveToOffset(extractedKey.getTextOffset());
-        editor.getScrollingModel().scrollToCaret(ScrollType.MAKE_VISIBLE);
+      var showOpenFileInEditor = extractRefDialog.shouldOpenFileInEditor();
+      if (showOpenFileInEditor) {
+        FileEditorManager.getInstance(project).openFile(virtualFile, true);
       }
-      return extracted;
     });
-
-    ApplicationManager.getApplication().invokeLater(() -> startInplaceRename(extractedReference, editor));
-  }
-
-  private void startInplaceRename(YAMLKeyValue extractedReference, Editor editor) {
-    if (extractedReference == null) {
-      return;
-      // TODO show error message if the key is null
-    }
-    var dataContext = DataManager.getInstance().getDataContext(editor.getComponent());
-    var renameHandler = RenameHandlerRegistry.getInstance()
-            .getRenameHandler(dataContext);
-
-    if (renameHandler != null) {
-      renameHandler.invoke(extractedReference.getProject(), new PsiElement[] { extractedReference }, dataContext);
-    }
   }
 
   public boolean isAvailable(PsiElement element) {
-    return element instanceof YAMLKeyValue;
+    var file = element.getContainingFile();
+    if (!(file instanceof YAMLFile)) {
+      return false;
+    }
+    var isYamlFile = CIAidProjectService.isValidGitlabCIYamlFile(file.getVirtualFile());
+    if (isYamlFile) {
+      var topLevelKeys = YAMLUtil.getTopLevelKeys((YAMLFile) element.getContainingFile());
+      var isYamlKeyValue = element instanceof YAMLKeyValue;
+      if (!(isYamlKeyValue)) {
+        return false;
+      }
+      var keyValue = (YAMLKeyValue) element;
+      var isTopLevelKey = topLevelKeys.stream()
+              .anyMatch(topLevelKeyValue -> topLevelKeyValue.getKeyText().equals(keyValue.getKeyText()));
+      return !isTopLevelKey;
+    }
+    return false;
   }
 }